requirements.txt
dlib==19.24.2
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.0
pillow==10.2.0
tk==0.1.0
tkcalendar==1.6.1
tkintertable==1.3.3
